// apps/enterprise-service-management/src/generated/models/business-rule-dependency-model.ts
import type { BusinessRuleVersion } from './business-rule-version-model';
import type { Workspace } from './workspace-model';

export const BusinessRuleDependencyDependencyTypeKeyToLabel = {
  'Field': 'Field',
  'Rule': 'Rule',
  'Service': 'Service',
  'Workflow': 'Workflow',
  'Lifecycle': 'Lifecycle',
  'Approval': 'Approval',
  'SLA': 'SLA',
  'Report': 'Report'
} as const;
export type BusinessRuleDependencyDependencyTypeKey = keyof typeof BusinessRuleDependencyDependencyTypeKeyToLabel;

export const BusinessRuleDependencyStatusKeyToLabel = {
  'Active': 'Active',
  'Inactive': 'Inactive'
} as const;
export type BusinessRuleDependencyStatusKey = keyof typeof BusinessRuleDependencyStatusKeyToLabel;

export interface BusinessRuleDependency {
  /**
   * @displayName Business Rule Dependency ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Display Label
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  displayLabel: string;
  /**
   * @displayName Business Rule Version
   * @validationRule Required for create/update operations
   */
  businessRuleVersion: Pick<BusinessRuleVersion, 'id' | 'snapshotLabel'>;
  /**
   * @displayName Created At
   * @validationRule Required for create/update operations
   */
  createdAt: string;
  /**
   * @displayName Dependency Type
   * @validationRule Required for create/update operations
   */
  dependencyTypeKey: BusinessRuleDependencyDependencyTypeKey;
  /**
   * @displayName Is Deleted
   * @validationRule Required for create/update operations
   */
  isDeleted: boolean;
  /**
   * @displayName Referenced Record ID or Key
   * @validationRule Required for create/update operations
   */
  referencedRecordIDOrKey: string;
  /**
   * @displayName Required
   * @validationRule Required for create/update operations
   */
  required: boolean;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: BusinessRuleDependencyStatusKey;
  /**
   * @displayName Workspace
   * @validationRule Required for create/update operations
   */
  workspace: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _BusinessRuleDependency = 'BusinessRuleDependency' as const;