// src/generated/models/business-rule-override-model.ts
import type { BusinessRuleDefinition } from './business-rule-definition-model';
import type { Person } from './person-model';
import type { FormVersion } from './form-version-model';
import type { Service } from './service-model';
import type { Workspace } from './workspace-model';

export const BusinessRuleOverrideLocalOverrideModeKeyToLabel = {
  'Inherit': 'Inherit',
  'Extend': 'Extend',
  'Replace': 'Replace',
  'Disable': 'Disable'
} as const;
export type BusinessRuleOverrideLocalOverrideModeKey = keyof typeof BusinessRuleOverrideLocalOverrideModeKeyToLabel;

export const BusinessRuleOverrideStatusKeyToLabel = {
  'Active': 'Active',
  'Inactive': 'Inactive'
} as const;
export type BusinessRuleOverrideStatusKey = keyof typeof BusinessRuleOverrideStatusKeyToLabel;

export interface BusinessRuleOverride {
  /**
   * @displayName Business Rule Override ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Reason
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  reason: string;
  /**
   * @displayName Business Rule Definition
   * @validationRule Required for create/update operations
   */
  businessRuleDefinition: Pick<BusinessRuleDefinition, 'id' | 'name1'>;
  /**
   * @displayName Created At
   * @validationRule Required for create/update operations
   */
  createdAt: string;
  /**
   * @displayName Created By
   * @validationRule Required for create/update operations
   */
  createdBy: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName Form Version
   */
  formVersion?: Pick<FormVersion, 'id' | 'snapshotLabel'>;
  /**
   * @displayName Is Deleted
   * @validationRule Required for create/update operations
   */
  isDeleted: boolean;
  /**
   * @displayName Local Override Mode
   * @validationRule Required for create/update operations
   */
  localOverrideModeKey: BusinessRuleOverrideLocalOverrideModeKey;
  /**
   * @displayName Override Condition
   */
  overrideCondition?: string;
  /**
   * @displayName Override Configuration
   */
  overrideConfiguration?: string;
  /**
   * @displayName Service
   */
  service?: Pick<Service, 'id' | 'serviceName'>;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: BusinessRuleOverrideStatusKey;
  /**
   * @displayName Updated At
   * @validationRule Required for create/update operations
   */
  updatedAt: string;
  /**
   * @displayName Updated By
   * @validationRule Required for create/update operations
   */
  updatedBy: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName Workspace
   * @validationRule Required for create/update operations
   */
  workspace: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _BusinessRuleOverride = 'BusinessRuleOverride' as const;