// apps/enterprise-service-management/src/generated/models/business-rule-usage-model.ts
import type { BusinessRuleVersion } from './business-rule-version-model';
import type { Workspace } from './workspace-model';

export const BusinessRuleUsageConsumerTypeKeyToLabel = {
  'Service': 'Service',
  'FormVersion': 'FormVersion',
  'FieldDefinition': 'FieldDefinition',
  'Workflow': 'Workflow',
  'Lifecycle': 'Lifecycle',
  'Approval': 'Approval',
  'SLA': 'SLA',
  'Report': 'Report'
} as const;
export type BusinessRuleUsageConsumerTypeKey = keyof typeof BusinessRuleUsageConsumerTypeKeyToLabel;

export const BusinessRuleUsageStatusKeyToLabel = {
  'Active': 'Active',
  'Inactive': 'Inactive'
} as const;
export type BusinessRuleUsageStatusKey = keyof typeof BusinessRuleUsageStatusKeyToLabel;

export interface BusinessRuleUsage {
  /**
   * @displayName Business Rule Usage ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Consumer Label
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  consumerLabel: string;
  /**
   * @displayName Business Rule Version
   * @validationRule Required for create/update operations
   */
  businessRuleVersion: Pick<BusinessRuleVersion, 'id' | 'snapshotLabel'>;
  /**
   * @displayName Consumer Record ID
   * @validationRule Required for create/update operations
   */
  consumerRecordID: string;
  /**
   * @displayName Consumer Type
   * @validationRule Required for create/update operations
   */
  consumerTypeKey: BusinessRuleUsageConsumerTypeKey;
  /**
   * @displayName Created At
   * @validationRule Required for create/update operations
   */
  createdAt: string;
  /**
   * @displayName Is Deleted
   * @validationRule Required for create/update operations
   */
  isDeleted: boolean;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: BusinessRuleUsageStatusKey;
  /**
   * @displayName Usage Context
   * @validationRule Required for create/update operations
   */
  usageContext: string;
  /**
   * @displayName Workspace
   * @validationRule Required for create/update operations
   */
  workspace: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _BusinessRuleUsage = 'BusinessRuleUsage' as const;