// apps/enterprise-service-management/src/generated/models/business-rule-version-model.ts
import type { BusinessRuleDefinition } from './business-rule-definition-model';
import type { Person } from './person-model';

export const BusinessRuleVersionStatusKeyToLabel = {
  'Draft': 'Draft',
  'Published': 'Published',
  'Retired': 'Retired'
} as const;
export type BusinessRuleVersionStatusKey = keyof typeof BusinessRuleVersionStatusKeyToLabel;

export interface BusinessRuleVersion {
  /**
   * @displayName Business Rule Version ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Snapshot Label
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  snapshotLabel: string;
  /**
   * @displayName Business Rule Definition
   * @validationRule Required for create/update operations
   */
  businessRuleDefinition: Pick<BusinessRuleDefinition, 'id' | 'name1'>;
  /**
   * @displayName Created At
   * @validationRule Required for create/update operations
   */
  createdAt: string;
  /**
   * @displayName Effective From
   */
  effectiveFrom?: string;
  /**
   * @displayName Effective To
   */
  effectiveTo?: string;
  /**
   * @displayName Immutable Condition Expression
   * @validationRule Required for create/update operations
   */
  immutableConditionExpression: string;
  /**
   * @displayName Immutable Outcome Configuration
   * @validationRule Required for create/update operations
   */
  immutableOutcomeConfiguration: string;
  /**
   * @displayName Is Deleted
   * @validationRule Required for create/update operations
   */
  isDeleted: boolean;
  /**
   * @displayName Published At
   */
  publishedAt?: string;
  /**
   * @displayName Published By
   */
  publishedBy?: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: BusinessRuleVersionStatusKey;
  /**
   * @displayName Updated At
   * @validationRule Required for create/update operations
   */
  updatedAt: string;
  /**
   * @displayName Version Number
   * @validationRule Required for create/update operations
   */
  versionNumber: number;
}

export const _BusinessRuleVersion = 'BusinessRuleVersion' as const;